+++ packages/grafana-ui/src/components/FormField/FormField.tsx
@@ -1,5 +1,5 @@
 import React, { InputHTMLAttributes, FunctionComponent } from 'react';
-import { FormLabel } from '../FormLabel/FormLabel';
+import { InlineFormLabel } from '../Forms/InlineFormLabel';
 import { PopoverContent } from '../Tooltip/Tooltip';
 import { cx } from 'emotion';
 
@@ -32,9 +32,9 @@ export const FormField: FunctionComponent<Props> = ({
 }) => {
   return (
     <div className={cx('form-field', className)}>
-      <FormLabel width={labelWidth} tooltip={tooltip}>
+      <InlineFormLabel width={labelWidth} tooltip={tooltip}>
         {label}
-      </FormLabel>
+      </InlineFormLabel>
       {inputEl || (
         <input type="text" className={`gf-form-input ${inputWidth ? `width-${inputWidth}` : ''}`} {...inputProps} />
       )}
